+++ bin/open-git-url.js
@@ -6,7 +6,11 @@ const {cliBasics} = require('command-line-basics');
 const mainScript = require('../src/index.js');
 
 const optionDefinitions = cliBasics(
-  join(__dirname, '../src/optionDefinitions.js')
+  join(__dirname, '../src/optionDefinitions.js'), {
+    commandLineArgsOptions: {
+      camelCase: true
+    }
+  }
 );
 
 if (!optionDefinitions) { // cliBasics handled